requirements.txt
langchain
langchain-experimental
langchainhub
langchain-community
langchain-text-splitters
langchain-openai
langsmith
langchain-core
langgraph
langchain-pinecone
requests
black
openai
python-dotenv
pinecone
beautifulsoup4
altair<5
firecrawl-py
qrcode
streamlit
pandas
tabulate
matplotlib